Fix math error in StelHips
This bug was probably making the allsky images useless at order different
than zero.

+++ src/core/StelHips.cpp
@@ -268,7 +268,7 @@ HipsTile* HipsSurvey::getTile(int order, int pix)
 		// Use the allsky image until we load the full texture.
 		if (order == orderMin && !allsky.isNull())
 		{
-			int nbw = sqrt(12 * 1 << (2 * order));
+			int nbw = (int)sqrt(12 * (1 << (2 * order)));
 			int x = (pix % nbw) * allsky.width() / nbw;
 			int y = (pix / nbw) * allsky.width() / nbw;
 			int s = allsky.width() / nbw;
